+++ subprojects/ide/src/main/java/org/gradle/plugins/ide/eclipse/model/ResourceFilterMatcher.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011 the original author or authors.
+ * Copyright 2017 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,19 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package org.gradle.plugins.ide.eclipse.model;
 
-import com.google.common.base.Objects;
-import com.google.common.collect.Sets;
-import groovy.lang.Closure;
-import groovy.lang.DelegatesTo;
 import org.gradle.api.Action;
-import org.gradle.api.InvalidUserDataException;
 import org.gradle.api.Nullable;
-import org.gradle.api.internal.ClosureBackedAction;
 
 import java.util.Set;
-
 /**
  * The model of an Eclipse resource filter matcher.
  * <p>
@@ -42,102 +36,44 @@ import java.util.Set;
  *
  * @since 3.5
  */
-public final class ResourceFilterMatcher {
-    private String id;
-    private String arguments;
-    private Set<ResourceFilterMatcher> children = Sets.newLinkedHashSet();
-
-    public ResourceFilterMatcher() {
-    }
-
-    public ResourceFilterMatcher(String id, String arguments, Set<ResourceFilterMatcher> children) {
-        this();
-        this.id = id;
-        this.arguments = arguments;
-        this.children = children;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
+public interface ResourceFilterMatcher {
+    /**
+     * The id of the matcher type, as defined by its Eclipse extension.
+     *
+     * Can be null on a newly created matcher, but must be set in order to be considered valid.
+     */
     @Nullable
-    public String getArguments() {
-        return arguments;
-    }
-
-    public void setArguments(String arguments) {
-        this.arguments = arguments;
-    }
-
-    public Set<ResourceFilterMatcher> getChildren() {
-        return children;
-    }
-
-    public void setChildren(Set<ResourceFilterMatcher> children) {
-        if (children == null) {
-            throw new InvalidUserDataException("children must not be null");
-        }
-        this.children = children;
-    }    
+    String getId();
 
     /**
-     * Adds a child/nested matcher to this matcher.
+     * Sets the id of the matcher type.
      *
-     * @param configureClosure The closure to use to configure the matcher.
+     * @param id the id, cannot be null
      */
-    public ResourceFilterMatcher matcher(@DelegatesTo(value=ResourceFilterMatcher.class, strategy = Closure.DELEGATE_FIRST) Closure configureClosure) {
-        return matcher(new ClosureBackedAction<ResourceFilterMatcher>(configureClosure));
-    }
+    void setId(String id);
+
+    /**
+     * The arguments of the matcher or null if it has children.
+     */
+    @Nullable
+    String getArguments();
+
+    /**
+     * Sets the arguments of the matcher.
+     *
+     * @param arguments the arguments or null if the matcher should have child matchers instead
+     */
+    void setArguments(String arguments);
+
+    /**
+     * The child matchers of this matcher, e.g. when this is an OR-matcher.
+     */
+    Set<ResourceFilterMatcher> getChildren();
 
     /**
      * Adds a child/nested matcher to this matcher.
      *
      * @param configureAction The action to use to configure the matcher.
      */
-    public ResourceFilterMatcher matcher(Action<? super ResourceFilterMatcher> configureAction) {
-        ResourceFilterMatcher m = new ResourceFilterMatcher();
-        configureAction.execute(m);
-        children.add(m);
-        return m;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null) {
-            return false;
-        }
-        if (!getClass().equals(o.getClass())) {
-            return false;
-        }
-        ResourceFilterMatcher resourceFilterMatcher = (ResourceFilterMatcher) o;
-        return Objects.equal(id, resourceFilterMatcher.id)
-            && Objects.equal(arguments, resourceFilterMatcher.arguments)
-            && Objects.equal(children, resourceFilterMatcher.children);
-    }
-
-    @Override
-    public int hashCode() {
-        int result;
-        result = id != null ? id.hashCode() : 0;
-        result = 31 * result + (arguments != null ? arguments.hashCode() : 0);
-        result = 31 * result + (children != null ? children.hashCode() : 0);
-        return result;
-    }
-
-    @Override
-    public String toString() {
-        return "ResourceFilterMatcher{"
-            + "id='" + id + '\''
-            + ", arguments='" + arguments + '\''
-            + ", children='" + children + '\''
-            + '}';
-    }    
+    ResourceFilterMatcher matcher(Action<? super ResourceFilterMatcher> configureAction);
 }
