+++ subprojects/ide/src/test/groovy/org/gradle/plugins/ide/eclipse/model/ProjectTest.groovy
@@ -17,6 +17,8 @@ package org.gradle.plugins.ide.eclipse.model;
 
 
 import org.gradle.internal.xml.XmlTransformer
+import org.gradle.plugins.ide.eclipse.model.internal.DefaultResourceFilter
+import org.gradle.plugins.ide.eclipse.model.internal.DefaultResourceFilterMatcher
 import org.gradle.test.fixtures.file.TestNameTestDirectoryProvider
 import org.junit.Rule
 import spock.lang.Specification
@@ -24,11 +26,11 @@ import spock.lang.Specification
 public class ProjectTest extends Specification {
     def static final CUSTOM_REFERENCED_PROJECTS = ['refProject'] as LinkedHashSet
     def static final CUSTOM_BUILD_COMMANDS = [new BuildCommand('org.eclipse.jdt.core.scalabuilder', [climate: 'cold'])]
-    def static final CUSTOM_NATURES = ['org.eclipse.jdt.core.scalanature'] 
+    def static final CUSTOM_NATURES = ['org.eclipse.jdt.core.scalanature']
     def static final CUSTOM_LINKED_RESOURCES = [new Link('somename', 'sometype', 'somelocation', '')] as Set
     def static final CUSTOM_RESOURCE_FILTERS = [
-        new ResourceFilter(ResourceFilterAppliesTo.FILES_AND_FOLDERS, ResourceFilterType.EXCLUDE_ALL, true, new ResourceFilterMatcher('org.eclipse.some.custom.matcher', 'foobar', [] as LinkedHashSet)),
-        new ResourceFilter(ResourceFilterAppliesTo.FOLDERS, ResourceFilterType.INCLUDE_ONLY, false, new ResourceFilterMatcher('org.eclipse.ui.ide.orFilterMatcher', null, [new ResourceFilterMatcher('org.eclipse.ui.ide.multiFilter', '1.0-name-matches-false-false-node_modules', [] as LinkedHashSet), new ResourceFilterMatcher('org.eclipse.ui.ide.multiFilter', '1.0-name-matches-false-false-target', [] as LinkedHashSet)] as LinkedHashSet))] as LinkedHashSet
+        new DefaultResourceFilter(ResourceFilterAppliesTo.FILES_AND_FOLDERS, ResourceFilterType.EXCLUDE_ALL, true, new DefaultResourceFilterMatcher('org.eclipse.some.custom.matcher', 'foobar', [] as LinkedHashSet)),
+        new DefaultResourceFilter(ResourceFilterAppliesTo.FOLDERS, ResourceFilterType.INCLUDE_ONLY, false, new DefaultResourceFilterMatcher('org.eclipse.ui.ide.orFilterMatcher', null, [new DefaultResourceFilterMatcher('org.eclipse.ui.ide.multiFilter', '1.0-name-matches-false-false-node_modules', [] as LinkedHashSet), new DefaultResourceFilterMatcher('org.eclipse.ui.ide.multiFilter', '1.0-name-matches-false-false-target', [] as LinkedHashSet)] as LinkedHashSet))] as LinkedHashSet
 
     @Rule
     public TestNameTestDirectoryProvider tmpDir = new TestNameTestDirectoryProvider();
@@ -57,8 +59,8 @@ public class ProjectTest extends Specification {
         eclipseProject.natures = ['constructorNature']
         eclipseProject.linkedResources = [new Link('constructorName', 'constructorType', 'constructorLocation', '')] as Set
         eclipseProject.resourceFilters = [
-            new ResourceFilter(ResourceFilterAppliesTo.FILES, ResourceFilterType.INCLUDE_ONLY, false, new ResourceFilterMatcher('matcherId', 'matcherArgs', [] as LinkedHashSet)),
-            new ResourceFilter(ResourceFilterAppliesTo.FILES_AND_FOLDERS, ResourceFilterType.EXCLUDE_ALL, true, new ResourceFilterMatcher('org.eclipse.ui.ide.orFilterMatcher', null, [new ResourceFilterMatcher('org.eclipse.ui.ide.multiFilter', '1.0-name-matches-false-false-node_modules', [] as LinkedHashSet)] as LinkedHashSet))] as LinkedHashSet
+            new DefaultResourceFilter(ResourceFilterAppliesTo.FILES, ResourceFilterType.INCLUDE_ONLY, false, new DefaultResourceFilterMatcher('matcherId', 'matcherArgs', [] as LinkedHashSet)),
+            new DefaultResourceFilter(ResourceFilterAppliesTo.FILES_AND_FOLDERS, ResourceFilterType.EXCLUDE_ALL, true, new DefaultResourceFilterMatcher('org.eclipse.ui.ide.orFilterMatcher', null, [new DefaultResourceFilterMatcher('org.eclipse.ui.ide.multiFilter', '1.0-name-matches-false-false-node_modules', [] as LinkedHashSet)] as LinkedHashSet))] as LinkedHashSet
 
         when:
         project.load(customProjectReader)
@@ -93,7 +95,7 @@ public class ProjectTest extends Specification {
         eclipseProject.name = 'constructorName'
         eclipseProject.comment = 'constructorComment'
         eclipseProject.referencedProjects = ['constructorRefProject'] as LinkedHashSet
-        eclipseProject.resourceFilters = [new ResourceFilter(ResourceFilterAppliesTo.FOLDERS, ResourceFilterType.INCLUDE_ONLY, true, new ResourceFilterMatcher('matcherId2', 'matcherArgs2', [] as LinkedHashSet))] as LinkedHashSet
+        eclipseProject.resourceFilters = [new DefaultResourceFilter(ResourceFilterAppliesTo.FOLDERS, ResourceFilterType.INCLUDE_ONLY, true, new DefaultResourceFilterMatcher('matcherId2', 'matcherArgs2', [] as LinkedHashSet))] as LinkedHashSet
 
         when:
         project.load(customProjectReader)
